fix: pastikan endpoint statistik mengembalikan integer

+++ app/Http/Controllers/Api/DashboardController.php
@@ -11,20 +11,20 @@ class DashboardController extends Controller
     public function stats()
     {
         return response()->json([
-            'total_aset' => Asset::sum('jumlah'),
-            'total_item' => Asset::count(),
-            'total_kategori' => Category::count(),
-            'total_lokasi' => Location::count(),
+            'total_aset' => (int) Asset::sum('jumlah'),
+            'total_item' => (int) Asset::count(),
+            'total_kategori' => (int) Category::count(),
+            'total_lokasi' => (int) Location::count(),
             'kondisi' => [
-                'baik' => Asset::where('kondisi', 'baik')->sum('jumlah'),
-                'rusak_ringan' => Asset::where('kondisi', 'rusak_ringan')->sum('jumlah'),
-                'rusak_berat' => Asset::where('kondisi', 'rusak_berat')->sum('jumlah'),
+                'baik' => (int) Asset::where('kondisi', 'baik')->sum('jumlah'),
+                'rusak_ringan' => (int) Asset::where('kondisi', 'rusak_ringan')->sum('jumlah'),
+                'rusak_berat' => (int) Asset::where('kondisi', 'rusak_berat')->sum('jumlah'),
             ],
             'aset_per_kategori' => Category::withSum('assets', 'jumlah')
                 ->get()
                 ->map(fn ($cat) => [
                     'nama_kategori' => $cat->nama_kategori,
-                    'total' => $cat->assets_sum_jumlah ?? 0,
+                    'total' => (int) $cat->assets_sum_jumlah ?? 0,
                 ]),
         ]);
     }
